refactor: Referee의 역할을 RacingGame이 가지도록 수정
