+++ src/main/java/racingcar/domain/RacingGame.java
@@ -8,11 +8,23 @@ public class RacingGame {
 
     private static final String CAR_NAME_DELIMITER = ",";
 
-    private final Referee referee;
+    private final int roundToPlay;
+    private int currentRound;
+    private final List<Car> registeredCars;
+    private final NumberSelector numberSelector;
 
     public RacingGame(String inputCarNames, int roundToPlay, NumberSelector numberSelector) {
-        List<Car> registeredCars = registerCars(inputCarNames);
-        this.referee = new Referee(roundToPlay, registeredCars, numberSelector);
+        validateRoundToPlay(roundToPlay);
+        this.roundToPlay = roundToPlay;
+        this.currentRound = 0;
+        this.registeredCars = registerCars(inputCarNames);
+        this.numberSelector = numberSelector;
+    }
+
+    private void validateRoundToPlay(int roundToPlay) {
+        if (roundToPlay < 0) {
+            throw new IllegalArgumentException("라운드는 양수여야 합니다.");
+        }
     }
 
     private List<Car> registerCars(String inputCarNames) {
@@ -26,22 +38,40 @@ public class RacingGame {
         if (isGamePlaying()) {
             throw new IllegalArgumentException("아직 게임이 종료되지 않았습니다.");
         }
-        return referee.announceWinners();
+        int maxPosition = calculateMaxCarPosition();
+        return registeredCars.stream()
+                .filter(car -> car.isSamePosition(maxPosition))
+                .collect(Collectors.toList());
     }
 
-    public boolean isGameEnded() {
-        return referee.isGameEnded();
-    }
-
-    public boolean isGamePlaying() {
-        return referee.isGamePlaying();
+    private int calculateMaxCarPosition() {
+        int maxPosition = 0;
+        for (Car car : registeredCars) {
+            maxPosition = Math.max(maxPosition, car.getPosition());
+        }
+        return maxPosition;
     }
 
     public void proceedRound() {
-        referee.moveCars();
+        if (isGameEnded()) {
+            throw new IllegalArgumentException("게임이 종료되었습니다.");
+        }
+        for (Car car : registeredCars) {
+            int number = numberSelector.selectNumber();
+            car.move(number);
+        }
+        currentRound++;
+    }
+
+    public boolean isGameEnded() {
+        return currentRound == roundToPlay;
+    }
+
+    public boolean isGamePlaying() {
+        return currentRound < roundToPlay;
     }
 
     public List<Car> announceRoundResult() {
-        return referee.announceRoundResult();
+        return registeredCars;
     }
 }
